consumer/requirements.txt
kafka-python
weaviate-client
feedparser
requests
beautifulsoup4
python-dateutil
sentence-transformers